all: support Go1.9 with type alias Buffer instead of strings.Builder
strings.Builder was introduced in Go1.10 but this package
uses it for efficiency.
Luckily we've got type aliases, build constraints as well
as in Go1.12, we have the gcflag "-lang" that we can
use to test out compiler versions.

This change adds a Buffer type in syntax, which is:
* type Buffer = bytes.Buffer on Go1.9 and below
* type Buffer = strings.Builder on Go1.10 and above

I verified the builds by doing
```shell
go test -v -gcflags="-lang=go1.9 ./...
go test -v -gcflags="-lang=go1.10 ./...
```

Fixes #1